refactored the tests to make them a bit more concise and clear. also increased test coverage on copier by adding checking of the output via a small interface change

+++ spec/bak/copier_spec.rb
@@ -1,47 +1,39 @@
 require 'spec_helper'
 require "rspec/expectations"
 
+
 describe "Copier" do
     before(:each) do
         @file = "testfile.txt"
         @generator = BackupNameGenerator.new(@file, {})
-        @copier = FileCopier.new(@file, @generator)
+        @output = double('stdout')
+        @copier = FileCopier.new(@generator, @output)
         FileUtils.touch(@file)
     end
 
-    before("file doesn't exist") do
-        FileUtils.rm(@file)
-    end
-
     after(:each) do
-        begin
-            FileUtils.rm("#{@file}*")
-        rescue Exception => e
-        end
+        FileUtils.rm("#{@file}.bak") if File.file?("#{@file}.bak")
+        FileUtils.rm(@file) if File.file?(@file)
     end
 
-    describe "public interface" do
-        it { @copier.start_file.should == @file }
-        it { @copier.end_file.should == "#{@file}.bak" }
-        it { @copier.should respond_to(:start) }
+    it { @copier.start_file.should == @file }
+    it { @copier.end_file.should == "#{@file}.bak" }
+    it { @copier.should respond_to(:start) }
+
+    it "should return nil to indicate that no errors were received from start" do
+        @copier.start.should be_nil
     end
 
-    describe "no errors" do
-        it { @copier.start.should be_nil }
+    it "should show a message indicating that the base file is missing" do
+        FileUtils.rm(@file) if File.file?(@file)
+        @output.should_receive(:puts).with("#{@file}: No such file or directory")
+        @copier.start
     end
 
-
-    #describe "file doesn't exist" do
-    #    it { STDERR.should_receive(:print).with("No such file or directory") }
-    #    it { @copier.start }
-    #end
-
-    #    #describe "target file already exists and no force setting" do
-    #    #    FileUtils.touch(@file)
-    #    #    FileUtils.touch("#{@file}.bak")
-    #    #    copier = FileCopier.new(@file, @generator)
-    #    #    STDERR.should_receive(:print).with("blah blah")
-    #    #    copier.start
-    #    #end
-    #end
+    it "should tell you if the target file already exists and force mode isn't on" do
+        FileUtils.touch(@file)
+        FileUtils.touch("#{@file}.bak")
+        @output.should_receive(:puts).with("#{@copier.end_file}: File already exists")
+        @copier.start
+    end
 end
